moved image controller with newly created image.service in image folder - moved all multer uploading code from post controller to image service
